successful addition of OL component
