--- src/Components/Button.js
@@ -1,12 +1,0 @@
-
-import React from 'react';
-
-function Button(props) {
-    return (
-       <button>
-        ADOPT
-       </button>
-    );
-}
-
-export default Button;
